Generate a bi-criteria graph with two uniformly randomly distribted edge weights.

The instance is composed of two symmetric weight matrices. The first weight is drawn independently at random from a \(\mathcal{R}[10, 100]\) distribution, the second one from a \(\mathcal{R}[10, 50]\) distribution (see references).

genRandomMCGP(n)

Arguments

n

[integer(1)]
Instance size, i.e., number of nodes.

Value

[grapherator] Graph.

Note

This is a simple wrapper around the much more flexible graph generation system in package grapherator.

References

Zhou, G. and Gen, M. Genetic Algorithm Approach on Multi-Criteria Minimum Spanning Tree Problem. In: European Journal of Operational Research (1999).

Knowles, JD & Corne, DW 2001, A comparison of encodings and algorithms for multiobjective minimum spanning tree problems. in Proceedings of the IEEE Conference on Evolutionary Computation, ICEC|Proc IEEE Conf Evol Comput Proc ICEC. vol. 1, Institute of Electrical and Electronics Engineers , pp. 544-551, Congress on Evolutionary Computation 2001, Soul, 1 July.

Examples

g = genRandomMCGP(10L)
if (FALSE) {
pl = grapherator::plot(g)
}