Multi-Objective Prim algorithm.

Approximates the Pareto-optimal mcMST front of a multi-objective graph problem by iteratively applying Prim's algorithm for the single-objective MST problem to a scalarized version of the problem. I.e., the weight vector \((w_1, w_2)\) of an edge \((i, j)\) is substituted with a weighted sum \(\lambda_i w_1 + (1 - \lambda_i) w_2\) for different weights \(\lambda_i \in [0, 1]\).

mcMSTPrim(instance, n.lambdas = NULL, lambdas = NULL)

Arguments

instance

[grapherator]
Graph.

n.lambdas

[integer(1) | NULL]
Number of weights to generate. The weights are generated equdistantly in the interval \([0, 1]\).

lambdas

[numerci]
Vector of weights. This is an alternative to n.lambdas.

Value

[list] List with component pareto.front.

Note

Note that this procedure can only find socalled supported efficient solutions, i.e., solutions on the convex hull of the Pareto-optimal front.

References

J. D. Knowles and D. W. Corne, "A comparison of encodings and algorithms for multiobjective minimum spanning tree problems," in Proceedings of the 2001 Congress on Evolutionary Computation (IEEE Cat. No.01TH8546), vol. 1, 2001, pp. 544–551 vol. 1.

See also

Examples

g = genRandomMCGP(30)
res = mcMSTPrim(g, n.lambdas = 50)
print(res$pareto.front)
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,] 1746 1634 1474 1330 1218 1097 1097  982  914   909   909   909   853   805
#> [2,]  365  366  372  380  389  401  401  418  430   431   431   431   449   465
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]   774   774   680   680   666   666   650   632   632   597   597   595
#> [2,]   477   477   520   520   528   528   539   552   552   581   581   583
#>      [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37] [,38]
#> [1,]   595   557   547   539   516   506   468   458   454   454   454   454
#> [2,]   583   628   641   652   686   702   769   788   797   797   797   797
#>      [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49] [,50]
#> [1,]   454   454   446   441   440   440   438   436   436   436   436   436
#> [2,]   797   797   830   854   860   860   876   895   895   895   895   895