Uniform mutation for Pruefer code representation.

mutUniformPruefer replaces each component of a Pruefer code of length n - 2 with probability p with a random node number between 1 and n.

mutUniformPruefer(ind, p = 1/length(ind))

Arguments

ind

[integer]
Pruefer code.

p

[numeric(1)]
Probability of mutation of each component of ind. Default is 1 / length(ind).

Value

[integer] Mutated Pruefer code.

See also