Visualization of edge frequency among solution set.

Given a list of graphs and a list of solutions (encoded as edge lists) for each graph the function generates each one plot. This is a 2D-scatterplot of edge weights of the graph. Size and colour of each point indicate the number of solutions the edge is part of.

plotEdgeFrequency(graphs, approx.sets, facet.args = list(), names = NULL)

Arguments

graphs

[list(grapherator)]
List of grapherator graphs.

approx.sets

[list(list(matrix))]
List of approximations sets.

facet.args

[list]
Further arguments passed down to facet_wrap. Only relevant if length(graphs) > 1.

names

[character]
Optional names of the graph instances. Used for facetting. Default is “Problem_i” with i ranging from 1 to length(graphs).

Value

[ggplot]

See also

Examples

g = genRandomMCGP(50L)
res = mcMSTEmoaBG(mu = 10L, max.iter = 50, instance = g, scalarize = TRUE)
if (FALSE) {
plotEdgeFrequency(list(g), list(res$pareto.set))
}