Multi-Objective Prim algorithm.

Approximates the Pareto-optimal mcMST front of a multi-objective graph problem by iteratively applying Prim's algorithm for the single-objective MST problem to a scalarized version of the problem. I.e., the weight vector \((w_1, w_2)\) of an edge \((i, j)\) is substituted with a weighted sum \(\lambda_i w_1 + (1 - \lambda_i) w_2\) for different weights \(\lambda_i \in [0, 1]\).

mcMSTPrim(instance, n.lambdas = NULL, lambdas = NULL)

Arguments

instance

[grapherator]
Graph.

n.lambdas

[integer(1) | NULL]
Number of weights to generate. The weights are generated equdistantly in the interval \([0, 1]\).

lambdas

[numerci]
Vector of weights. This is an alternative to n.lambdas.

Value

[list] List with component pareto.front.

Note

Note that this procedure can only find socalled supported efficient solutions, i.e., solutions on the convex hull of the Pareto-optimal front.

References

J. D. Knowles and D. W. Corne, "A comparison of encodings and algorithms for multiobjective minimum spanning tree problems," in Proceedings of the 2001 Congress on Evolutionary Computation (IEEE Cat. No.01TH8546), vol. 1, 2001, pp. 544–551 vol. 1.

See also

Examples

g = genRandomMCGP(30)
res = mcMSTPrim(g, n.lambdas = 50)
print(res$pareto.front)
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,] 1626 1562 1489 1284 1137 1057  999  966  950   869   844   811   766   725
#> [2,]  368  368  370  382  394  402  409  414  417   434   440   449   463   477
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]   640   633   606   592   592   592   583   583   553   545   532   532
#> [2,]   509   512   525   532   532   532   538   538   561   568   580   580
#>      [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37] [,38]
#> [1,]   521   521   492   483   483   483   476   476   464   464   453   453
#> [2,]   592   592   630   643   643   643   656   656   681   681   709   709
#>      [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49] [,50]
#> [1,]   453   449   439   437   437   433   433   433   433   431   431   431
#> [2,]   709   723   764   773   773   801   801   801   801   838   838   838